+++ src/applications/health-care-questionnaire/questionnaire/containers/QuestionnaireWrapper.jsx
@@ -9,7 +9,10 @@ import {
 import LoadingIndicator from '@department-of-veterans-affairs/component-library/LoadingIndicator';
 import App from './App';
 import BreadCrumbs from '../components/bread-crumbs/BreadCrumbs';
-import { getCurrentAppointmentId, setCurrentAppointmentId } from '../utils';
+import {
+  getCurrentAppointmentId,
+  setCurrentAppointmentId,
+} from '../../shared/utils';
 
 const QuestionnaireWrapper = ({
   location,
