+++ src/applications/health-care-questionnaire/shared/tests/appointment.id.from.url.unit.spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 
-import { getAppointmentIdFromUrl } from '../../../utils';
+import { getAppointmentIdFromUrl } from '../utils';
 
 describe('health care questionnaire -- utils -- get id from url in window', () => {
   it('window is undefined', () => {
