+++ src/applications/health-care-questionnaire/shared/tests/get.current.appointment.unit.spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 
-import { getCurrentAppointmentId } from '../../../utils';
+import { getCurrentAppointmentId } from '../utils';
 
 describe('health care questionnaire -- utils -- get current id from url or window', () => {
   it('window is undefined', () => {
