+++ src/applications/health-care-questionnaire/shared/utils/appointment.utils.js
