+++ src/applications/health-care-questionnaire/shared/utils/index.js
