+++ src/applications/health-care-questionnaire/shared/utils/session.utils.js
